requirements.txt
appdirs==1.4.4
APScheduler==3.7.0
blinker==1.4
click==7.1.2
Flask==1.1.2
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.5.1
greenlet==1.0.0
gunicorn==20.1.0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mypy-extensions==0.4.3
pathspec==0.8.1
psycopg2==2.8.6
PyJWT==2.0.1
python-dateutil==2.8.1
pytz==2021.1
regex==2021.4.4
selenium==3.141.0
six==1.15.0
SQLAlchemy==1.4.11
toml==0.10.2
tzlocal==2.1
urllib3==1.26.4
Werkzeug==1.0.1
